refactor(agent): split api_server/response_actions.rs into execute/, rollback/, ...
Convert the single 2,838-line response_actions module into a directory
of focused sub-modules, all under 600 lines, matching the Phase 2 plan
in the swarm architecture synthesis:

  response_actions/
    mod.rs                  re-exports + glob from parent scope
    validate.rs             actor identity, ttl, reason validation
    dispatch.rs             execute_edr_response_action switch
    failure.rs              sanitize, redact, record-failure
    persistence.rs          ledger append, receipt emit, transition IDs
    network_extension.rs    NE egress policy reload coordination
    egress_match.rs         egress target normalize + active-match lookup
    broker_revocation.rs    broker capability + integration secret revoke
    process_signal.rs       signal target validation + unix kill helpers
    execute/                per-action live execution
      collect_evidence.rs
      restrict_egress.rs
      quarantine_file.rs
      disable_persistence.rs
      revoke_grant.rs
      suspend_process_tree.rs
    rollback/               per-action rollback + TTL expiration
      restrict_egress.rs
      quarantine_file.rs
      disable_persistence.rs
      suspend_process_tree.rs
      expiration.rs

The pub(crate) surface is unchanged: api_server.rs continues to do
`pub(crate) use response_actions::*;`, and mod.rs aggregates each
sub-module the same way. External callers (api_server.rs, edr/handlers,
api_server tests, policy_history, receipts, policy_delta) build without
modification. Existing tests pass; the response_actions unit test for
validate_response_action_actor continues to resolve through the same
glob path.

Path-bounded persistence checks are imported from the
clawdstrike-fs-policy-paths crate extracted in Phase 1.
